(print_numbers): Use 'double' for loop index, not 'int', to avoid problems with integer overflow.  On almost all machines 'double' works in every case where 'int' works, and it works on other cases besides.

+++ src/seq.c
@@ -181,7 +181,7 @@ valid_format (const char *fmt)
 static void
 print_numbers (const char *fmt)
 {
-  int i;
+  double i;
 
   for (i = 0; /* empty */; i++)
     {
